First public release of tj_log.  Has not been incorporated into other components yet, not 100% sure that's best as they'll then be less self-contained.
Provides a simple library for managing a list of debug output
channels.  Has built-in channels for stdout, an sqlite database, and
logcat on Android.  The logcat output is not tested.  The sqlite
database might in the future have a run counter associated with log
messages, in addition to the timestamp.